lots of assess changes + slides

+++ sim/Sample.py
@@ -50,7 +50,7 @@ def collate(batch):
                 x[k] = [torch.from_numpy(v)]
 
     # convert to (single) tensors
-    y = torch.from_numpy(np.asarray(y)).long()
+    y = torch.from_numpy(np.asarray(y)).float()
     x = {k: torch.stack(v).float() for k, v in x.items()}
 
     # output is dictionary of tensors
